-refactored to move main logic to service method
