+++ app/server/appsmith-server/src/main/java/com/appsmith/server/controllers/ApplicationController.java
@@ -4,6 +4,7 @@ import com.appsmith.server.constants.Url;
 import com.appsmith.server.controllers.ce.ApplicationControllerCE;
 import com.appsmith.server.services.ApplicationPageService;
 import com.appsmith.server.services.ApplicationService;
+import com.appsmith.server.services.ApplicationSnapshotService;
 import com.appsmith.server.services.ThemeService;
 import com.appsmith.server.solutions.ApplicationFetcher;
 import com.appsmith.server.solutions.ApplicationForkingService;
@@ -20,10 +21,11 @@ public class ApplicationController extends ApplicationControllerCE {
                                  ApplicationFetcher applicationFetcher,
                                  ApplicationForkingService applicationForkingService,
                                  ImportExportApplicationService importExportApplicationService,
-                                 ThemeService themeService) {
+                                 ThemeService themeService,
+                                 ApplicationSnapshotService applicationSnapshotService) {
 
         super(service, applicationPageService, applicationFetcher, applicationForkingService,
-                importExportApplicationService, themeService);
+                importExportApplicationService, themeService, applicationSnapshotService);
 
     }
 }
