+++ app/server/appsmith-server/src/main/java/com/appsmith/server/services/ApplicationSnapshotServiceImpl.java
@@ -2,13 +2,18 @@ package com.appsmith.server.services;
 
 import com.appsmith.server.repositories.ApplicationSnapshotRepository;
 import com.appsmith.server.services.ce.ApplicationSnapshotServiceCEImpl;
+import com.appsmith.server.solutions.ApplicationPermission;
+import com.appsmith.server.solutions.ImportExportApplicationService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 @Slf4j
 @Service
 public class ApplicationSnapshotServiceImpl extends ApplicationSnapshotServiceCEImpl implements ApplicationSnapshotService {
-    public ApplicationSnapshotServiceImpl(ApplicationSnapshotRepository applicationSnapshotRepository) {
-        super(applicationSnapshotRepository);
+    public ApplicationSnapshotServiceImpl(ApplicationSnapshotRepository applicationSnapshotRepository,
+                                          ApplicationService applicationService,
+                                          ImportExportApplicationService importExportApplicationService,
+                                          ApplicationPermission applicationPermission) {
+        super(applicationSnapshotRepository, applicationService, importExportApplicationService, applicationPermission);
     }
 }
