+++ app/server/appsmith-server/src/main/java/com/appsmith/server/services/ce/ApplicationSnapshotServiceCE.java
@@ -1,10 +1,14 @@
 package com.appsmith.server.services.ce;
 
-import com.appsmith.server.domains.ApplicationSnapshot;
-import com.appsmith.server.dtos.ApplicationJson;
 import reactor.core.publisher.Mono;
 
 public interface ApplicationSnapshotServiceCE {
-
-    Mono<ApplicationSnapshot> createSnapshotForApplication(String applicationId, ApplicationJson applicationJson);
+    /**
+     * This method will create a new snapshot of the provided applicationId and branch name and store in the
+     * ApplicationSnapshot collection.
+     * @param applicationId ID of the application, default application ID if application is connected to Git
+     * @param branchName name of the Git branch, null or empty if not connected to Git
+     * @return Created snapshot ID
+     */
+    Mono<String> createApplicationSnapshot(String applicationId, String branchName);
 }
