+++ app/server/appsmith-server/src/main/java/com/appsmith/server/solutions/ce/ImportExportApplicationServiceCEImplV2.java
@@ -2247,20 +2247,4 @@ public class ImportExportApplicationServiceCEImplV2 implements ImportExportAppli
                     return analyticsService.sendObjectEvent(event, application, data);
                 });
     }
-
-    @Override
-    public Mono<String> createApplicationSnapshot(String applicationId, String branchName) {
-        return applicationService.findBranchedApplicationId(branchName, applicationId, applicationPermission.getEditPermission())
-                /* SerialiseApplicationObjective=VERSION_CONTROL because this API can be invoked from developers.
-                exportApplicationById method check for MANAGE_PERMISSION if SerialiseApplicationObjective=SHARE.
-                */
-                .flatMap(branchedAppId ->
-                    Mono.zip(
-                            exportApplicationById(branchedAppId, SerialiseApplicationObjective.VERSION_CONTROL),
-                            Mono.just(branchedAppId)
-                    )
-                )
-                .flatMap(objects -> applicationSnapshotService.createSnapshotForApplication(objects.getT2(), objects.getT1()))
-                .map(BaseDomain::getId);
-    }
 }
