+++ app/server/appsmith-server/src/test/java/com/appsmith/server/services/ApplicationSnapshotServiceTest.java
@@ -1,0 +1,9 @@
+package com.appsmith.server.services;
+
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+public class ApplicationSnapshotServiceTest {
+
+
+}
